refactor: move api-key view package outside server
